requirements.txt
aniso8601==3.0.2
Flask-Cors==3.0.6
asn1crypto==0.24.0
astroid==2.0.1
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
cryptography==2.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.12.0
flask-restplus==0.11.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
psycopg2-binary==2.7.5
py==1.5.4
pycparser==2.18
pylint==2.0.0
pytest==3.6.3
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11